requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
gunicorn==22.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1

